requirements.txt
flask>=2.0.0
gunicorn>=20.0.0
supabase>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
numpy>=1.20.0
pytz>=2021.1
python-dateutil>=2.8.1
beautifulsoup4>=4.12.0
pandas>=1.3.0